bitelearn/myapp/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.6.1
Markdown==3.6
mongoengine==0.28.2
pillow==10.4.0
pymongo==4.8.0
sqlparse==0.5.1
tzdata==2024.1